testing: test-viewer now renders static reports instead of running tests in the browser.
