+++ scripts/collect-comment-tests.ts
@@ -1,6 +1,11 @@
 import { watch } from "fs";
 import { parseArgs } from "util";
-import { generateReport } from "../src/test/reporting";
+import { generateReport, type TestReport, type TestResults } from "../src/test/reporting";
+import { collectTests } from "../src/test/commentTestCollector";
+import type { TestFunction } from "../src/test/commentTestTypes";
+import { buildSimpleCFG, requirementTests } from "../src/test/commentTestHandlers";
+import { graphToDot } from "../src/control-flow/render";
+import { formatSnapshotName, loadDOTSnapshots } from "./generate-dot-snapshots";
 
 const watchDir = import.meta.dir + "/../src";
 
@@ -16,9 +21,54 @@ const { values } = parseArgs({
   allowPositionals: true,
 });
 
+
+function runTestsFor(testFunc: TestFunction): TestResults[] {
+  const testResults: TestResults[] = [];
+  for (const [key, value] of Object.entries(testFunc.reqs)) {
+    const reqHandler = requirementTests[key];
+    if (!reqHandler) {
+      continue;
+    }
+    testResults.push({
+      reqName: key,
+      reqValue: value,
+      failure: reqHandler(testFunc),
+    });
+  }
+  return testResults;
+}
+
+function generateDOTFor(testFunc: TestFunction): string {
+  const cfg = buildSimpleCFG(testFunc.language, testFunc.function);
+  const dot = graphToDot(cfg);
+  return dot;
+}
+
+
+
+export async function generateReport(tests: TestFunction[]): Promise<TestReport[]> {
+  const testReports: TestReport[] = [];
+  const dotSnapshots = await loadDOTSnapshots();
+  for (const testFunc of tests) {
+    const results = runTestsFor(testFunc);
+    const dot = {
+      current: generateDOTFor(testFunc),
+      snapshot: dotSnapshots[formatSnapshotName(testFunc)] as string
+    };
+    const failed = results.some(result => result.failure) || dot.current !== dot.snapshot;
+    const name = formatSnapshotName(testFunc);
+    const code = testFunc.function.text
+    testReports.push({ dot, results, failed, name, code })
+  }
+  return testReports;
+}
+
+
+
+
 async function writeReport() {
-  const report = await generateReport();
-  Bun.write("./dist/tests/testReport.json", JSON.stringify(report));
+  const report = await generateReport(await collectTests());
+  await Bun.write("./dist/tests/testReport.json", JSON.stringify(report));
 }
 
 async function logAndContinue(fn: () => Promise<void>): Promise<void> {
